+++ test/test.cpp
@@ -23,8 +23,8 @@
 // }
 TEST(Utility, DegNormalization)
 {
-    double t1 = 361, t2 = -359, t3 = 721, t4 = -719;
-    std::vector<double> one{t1, t2, t3, t4};
+    float t1 = 361, t2 = -359, t3 = 721, t4 = -719;
+    std::vector<float> one{t1, t2, t3, t4};
     int expect, result;
     for (const auto &number : one)
     {
@@ -35,8 +35,8 @@ TEST(Utility, DegNormalization)
 }
 TEST(Utility, RadNormalization)
 {
-    double t1 = 361, t2 = -359, t3 = 721, t4 = -719;
-    std::vector<double> one{t1, t2, t3, t4};
+    float t1 = 361, t2 = -359, t3 = 721, t4 = -719;
+    std::vector<float> one{t1, t2, t3, t4};
     int expect, result;
     for (const auto &number : one)
     {
@@ -47,15 +47,15 @@ TEST(Utility, RadNormalization)
 }
 TEST(Utility, IsIntersect)
 {
-    Eigen::Vector2d p1(0, 0), p2(60, 0), p3(60, 30), p4(0, 30), p5(23.686, 10.8453);
-    std::vector<Eigen::Vector2d> polygon{p1, p2, p3, p4, p1};
-    std::vector<Eigen::Vector2d> intersect{p5};
-    // std::vector<Eigen::Vector2d> outside{p5, p7, p8};
+    Eigen::Vector2f p1(0, 0), p2(60, 0), p3(60, 30), p4(0, 30), p5(23.686, 10.8453);
+    std::vector<Eigen::Vector2f> polygon{p1, p2, p3, p4, p1};
+    std::vector<Eigen::Vector2f> intersect{p5};
+    // std::vector<Eigen::Vector2f> outside{p5, p7, p8};
     int expect, result;
     for (const auto &point : intersect)
     {
         expect = 1;
-        Eigen::Vector2d p9(10000, point.y());
+        Eigen::Vector2f p9(10000, point.y());
         result = Utility::IsIntersect(point, p9, polygon[1], polygon[2]);
         EXPECT_EQ(expect, result);
     }
@@ -63,17 +63,17 @@ TEST(Utility, IsIntersect)
 
 TEST(Utility, FindIntersectionPoint)
 {
-    Eigen::Vector2d p1(0, 0), p2(1, 1), p3(0, 1), p4(1, 0);
-    Eigen::Vector2d expect(1, 1);
-    Eigen::Vector2d result = Utility::FindIntersectionPoint(p1, p2, p3, p2);
+    Eigen::Vector2f p1(0, 0), p2(1, 1), p3(0, 1), p4(1, 0);
+    Eigen::Vector2f expect(1, 1);
+    Eigen::Vector2f result = Utility::FindIntersectionPoint(p1, p2, p3, p2);
     EXPECT_EQ(expect, result);
 }
 
 TEST(Utility, OnSegment)
 {
-    Eigen::Vector2d p1(0, 0), p2(1, 1), p3(0, 1), p4(1, 0), p5(0, 5), p6(0.5, 0), p7(-10, 1), p8(10, 1);
-    std::vector<Eigen::Vector2d> polygon{p1, p2, p3, p4, p1};
-    std::vector<Eigen::Vector2d> outside{p5, p7, p8};
+    Eigen::Vector2f p1(0, 0), p2(1, 1), p3(0, 1), p4(1, 0), p5(0, 5), p6(0.5, 0), p7(-10, 1), p8(10, 1);
+    std::vector<Eigen::Vector2f> polygon{p1, p2, p3, p4, p1};
+    std::vector<Eigen::Vector2f> outside{p5, p7, p8};
     bool expect, result;
     for (const auto &point : outside)
     {
@@ -88,7 +88,7 @@ TEST(Utility, OnSegment)
 
 TEST(Utility, IsAboveSegment)
 {
-    Eigen::Vector2d p1(0, 0), p2(1, 1), p3(0, 1), p4(1, 0);
+    Eigen::Vector2f p1(0, 0), p2(1, 1), p3(0, 1), p4(1, 0);
     int expect = 1;
     int result = Utility::IsAboveSegment(p1, p2, p3);
     EXPECT_EQ(expect, result);
@@ -99,10 +99,10 @@ TEST(Utility, IsAboveSegment)
 
 TEST(Utility, IsInsidePolygon)
 {
-    Eigen::Vector2d p1(0, 0), p2(60, 0), p3(60, 30), p4(0, 30), p5(23.686, 10.8453);
-    std::vector<Eigen::Vector2d> polygon{p1, p2, p3, p4, p1};
-    std::vector<Eigen::Vector2d> inside{p5};
-    // std::vector<Eigen::Vector2d> outside{p5, p7, p8};
+    Eigen::Vector2f p1(0, 0), p2(60, 0), p3(60, 30), p4(0, 30), p5(23.686, 10.8453);
+    std::vector<Eigen::Vector2f> polygon{p1, p2, p3, p4, p1};
+    std::vector<Eigen::Vector2f> inside{p5};
+    // std::vector<Eigen::Vector2f> outside{p5, p7, p8};
     int expect, result;
     for (const auto &point : inside)
     {
